refactor form component
