+++ app-web/src/components/Form/TopicForm.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useState } from 'react';
+import React, { Fragment, useState, useMemo } from 'react';
 import {
   TextInput,
   SelectDropdown,
@@ -7,6 +7,7 @@ import {
   StyledErrorMessage,
 } from './form';
 import { Form } from 'react-final-form';
+import { graphql, useStaticQuery } from 'gatsby';
 import arrayMutators from 'final-form-arrays';
 import { FieldArray } from 'react-final-form-arrays';
 import axios from 'axios';
@@ -14,8 +15,45 @@ import StyledButton from '../UI/Button/Button';
 import Loading from '../UI/Loading/Loading';
 import { useKeycloak } from '@react-keycloak/web';
 import { DEVHUB_API_URL } from '../../constants/api';
+import { flattenGatsbyGraphQL } from '../../utils/dataHelpers';
 
-export const TopicForm = () => {
+export const TopicForm = ({ operation }) => {
+  let config = {};
+
+  // Configuration for create a topic
+
+  if (operation === 'create') {
+    config = {
+      apiEndPoint: '/v1/topics/',
+      initialValue: {
+        sources: [
+          {
+            sourceType: null,
+            resourceType: null,
+          },
+        ],
+      },
+    };
+  }
+
+  // Configuration for edit a topic
+
+  if (operation === 'edit') {
+    const topicName = location.pathname
+      .split('/')
+      .pop()
+      .replace(/-/g, ' ');
+    const topicData = dataQuery(topicName);
+    config = {
+      apiEndPoint: '/v1/edit/topic/',
+      initialValue: {
+        topicName: topicData.name,
+        topicDescription: topicData.description,
+        sources: topicData.sourceProperties.sources,
+      },
+    };
+  }
+
   const [loading, setLoading] = useState(false);
 
   const [response, setResponse] = useState({ status: '', data: '' });
@@ -23,11 +61,12 @@ export const TopicForm = () => {
   const [showMessage, setShowMessage] = useState(false);
 
   const [keycloak] = useKeycloak();
+
   const onSubmit = async values => {
     setLoading(true);
-    values = convertToRegistryFormat(values);
+    values = convertToRegistryFormat(values, operation);
     try {
-      const res = await axios.post(`${DEVHUB_API_URL}/v1/topics/`, values, {
+      const res = await axios.post(`${DEVHUB_API_URL}${config.apiEndPoint}`, values, {
         headers: {
           Authorization: `Bearer ${keycloak.token}`,
         },
@@ -60,22 +99,13 @@ export const TopicForm = () => {
     }
   };
 
-  const initialValue = {
-    sources: [
-      {
-        sourceType: null,
-        resourceType: null,
-      },
-    ],
-  };
-
   return (
     <StylesWrapper>
       {loading && <Loading message="Loading ..." />}
       <Form
         onSubmit={onSubmit}
         mutators={{ ...arrayMutators }}
-        initialValues={initialValue}
+        initialValues={config.initialValue}
         render={({
           handleSubmit,
           form: {
@@ -140,14 +170,14 @@ export const TopicForm = () => {
   );
 };
 
-const convertToRegistryFormat = values => {
+const convertToRegistryFormat = (values, operation) => {
   const convertedFields = {
     name: values.topicName,
     description: values.topicDescription,
     sourceProperties: {
       sources: values.sources.map(source => ({
         sourceType: source.sourceType,
-        sourceProperties: getSourceProps(source),
+        sourceProperties: getSourceProps(source, operation),
         resourceType: source.resourceType,
       })),
     },
@@ -155,16 +185,18 @@ const convertToRegistryFormat = values => {
   return convertedFields;
 };
 
-const getSourceProps = source => {
+const getSourceProps = (source, operation) => {
   const properties = {};
-  properties.url = source.url;
+  properties.url = source.sourceProperties.url;
   if (source.sourceType === 'github') {
-    properties.owner = source.owner;
-    properties.repo = source.repo;
-    properties.files = [source.file];
+    properties.owner = source.sourceProperties.owner;
+    properties.repo = source.sourceProperties.repo;
+    console.log(operation);
+    properties.files =
+      operation === 'create' ? [source.sourceProperties.files] : source.sourceProperties.files;
   } else if (source.sourceType === 'web') {
-    properties.title = source.title;
-    properties.description = source.description;
+    properties.title = source.sourceProperties.title;
+    properties.description = source.sourceProperties.description;
   }
   return properties;
 };
@@ -173,30 +205,70 @@ const SubForm = (sourceType, name) => {
   if (sourceType === 'web') {
     return (
       <Fragment>
-        <TextInput label="Enter the source URL" name={`${name}.url`}></TextInput>
-        <TextInput label="Provide a title for your source" name={`${name}.title`}></TextInput>
+        <TextInput label="Enter the source URL" name={`${name}.sourceProperties.url`}></TextInput>
+        <TextInput
+          label="Provide a title for your source"
+          name={`${name}.sourceProperties.title`}
+        ></TextInput>
         <TextInput
           label="Describe your source in less than 140 characters"
-          name={`${name}.description`}
+          name={`${name}.sourceProperties.description`}
         ></TextInput>
       </Fragment>
     );
   } else if (sourceType === 'github') {
     return (
       <Fragment>
-        <TextInput label="Github Repository URL" name={`${name}.url`}></TextInput>
+        <TextInput label="Github Repository URL" name={`${name}.sourceProperties.url`}></TextInput>
         <TextInput
           label="Github Repositiry owner's github user name"
-          name={`${name}.owner`}
+          name={`${name}.sourceProperties.owner`}
         ></TextInput>
-        <TextInput label="Repository name" name={`${name}.repo`}></TextInput>
+        <TextInput label="Repository name" name={`${name}.sourceProperties.repo`}></TextInput>
         <TextInput
           label="Enter path to files to from root of your repository"
-          name={`${name}.file`}
+          name={`${name}.sourceProperties.files`}
         ></TextInput>
       </Fragment>
     );
   } else return null;
+};
+
+const dataQuery = topicName => {
+  const { topics } = useStaticQuery(graphql`
+    query {
+      topics: allTopicRegistryJson {
+        edges {
+          node {
+            name
+            description
+            sourceProperties {
+              sources {
+                sourceType
+                sourceProperties {
+                  url
+                  title
+                  description
+                  files
+                  owner
+                  repo
+                }
+              }
+            }
+            resourceType
+          }
+        }
+      }
+    }
+  `);
+  const topicsNode = useMemo(() => flattenGatsbyGraphQL(topics.edges), [topics.edges]);
+  let topicData = {};
+  topicsNode.map(topic => {
+    if (topic.name === topicName) {
+      topicData = topic;
+    }
+  });
+  return topicData;
 };
 
 export default TopicForm;
