+++ app-web/src/pages/editTopic.js
@@ -6,26 +6,26 @@ import { useKeycloak } from '@react-keycloak/web';
 import { SEO } from '../components/SEO/SEO';
 import { Title } from '../components/Page';
 import Main from '../components/Page/Main';
-import ExistingResourceForm from '../components/Form/ExistingResourceForm';
+import TopicForm from '../components/Form/TopicForm';
 
 export const editTopic = () => {
   // eslint-disable-next-line
   const [keycloak] = useKeycloak();
-  const isAuthenticated = true; //keycloak && keycloak.authenticated;
+  const isAuthenticated = keycloak && keycloak.authenticated;
   return (
     <Layout>
-      <SEO title="Edit topic" />
       <Main>
         <Title
           title="Content Contribution"
           subtitle={!isAuthenticated ? 'Login via IDIR or Github to continue' : 'Edit a topic'}
         />
         {isAuthenticated ? (
-          <ExistingResourceForm></ExistingResourceForm>
+          <TopicForm operation="edit"></TopicForm>
         ) : (
           <h4 css={{ color: 'red' }}>Not Authorized</h4>
         )}
       </Main>
+      <SEO title="Edit topic" />
     </Layout>
   );
 };
